Some Federated Code
Mostly just moving code around with this. Introduced a wrapper class
(which still needs to be tested) for splitting datasets for use in the
federated context. Should not interfere with the centralised code
whatsoever.

The new federated dataset should allow for all the types of data
distribution that we want to do

Implemented the Federated Client code. Still untested since I need to
finish the Server first. Big tests to do, fun times ahead.

Formatted and fixed a lot of files using
```
ruff check . --fix && ruff format .
```